SILGen: Keep the 'let x: NSFoo? = NSFoo()' failing initializer hack from being optimized.
If we see that we're injecting the result of an @objc initializer into an optional, perform the T -> T? conversion using unchecked_ref_bit_cast, which is opaque to the optimizer, instead of doing a proper optional injection. This prevents the optimizer from defeating our workaround for failing initializers. <rdar://problem/17941601>

Swift SVN r21096