Fix issue with switch_enum simplification.
I hit this building with another change that I have.

In cases where we have a switch_enum that is dominated by another
switch_enum, where the dominating switch_enum branches to a block with
no block argument, but the dominated switch_enum branches to a block
with a block argument, we need to synthesize the block argument to pass
by using unchecked_enum_data.

Swift SVN r21167