Fix crash when a block has two incoming edges from the same terminator.
I noticed this while reviewing some code. It's possible to have a single
conditional branch where both the true and false destination are the
same block. The argument simplification code would crash in that case
due to removeArgumentFromTerminator assuming this couldn't happen, and
because the loop over blocks would attempt to visit the same terminator
twice, with the first visit invalidating the iterator over
SILSuccessors.

Swift SVN r21245