Fix a problem found by ASAN.
Adding to a DenseMap can result in reallocating and moving the data in
the map. Because of this, the result of getDFSInfo() could not be used
across a call to addDFSInfo() since we were storing the DFS info
directly in the map.

This change allocates the DFS info on the heap to work around the issue.

Swift SVN r21454