[sil-combine] Simplify and generalize switch_enum_addr of loadable type promotion to switch_enum.
Previously we were attempting to look for a store before the switch_enum_addr
and were performing some sort of simplistic store forwarding. This is
unnecessary and causes us to miss potential opportunities to promote
switch_enum_addr of loadable types.

With this patch we always promote switch_enum_addr of loadable types to a load +
switch_enum. This ensures we canonicalize all switch_enum_addr of loadable types
to switch_enum and leaves the load-store forwarding issue to mem2reg and load
store forwarding.

Swift SVN r21479