Check @objc on properties as part of validation, not just full type checking.
Without this, we don't treat properties in other files as @objc or dynamic
even when they really are.

As part of this, stop trying to recover when an IBOutlet has the wrong type;
it just leads to spurious diagnostics when the outlet is used. We don't have
to worry about DI diagnostics because we don't even get there.

rdar://problem/18111806

Swift SVN r21501