Instead of passing type metadata to swift_initClassMetadata, pass the size and alignment of each field.  Take advantage of this to pass a constant size and alignment when possible.
This avoids the need to recursively find type metadata for
every field type, allowing generic recursively-structured
classes to be built.  There are a number of more complicated
cases that this approach isn't good enough for, but this
is good enough for now to fix rdar://18067671.

Also make an effort to properly support generic subclasses
of Objective-C classes.

Swift SVN r21506