fix <rdar://problem/15631987> QoI: DI emits two errors for the same problem at the same location
This is straight-forward, we just unique based on SourceLoc instead of by SILLocation.  In cases
where there are multiple SIL Instructions that map to the same file/line/column, the SourceLoc is
a better approximation of what the user cares about.

This wouldn't be super important, except that not having it causes a QoI regression 
in a patch I'm working on.



Swift SVN r21619