Enable this test now, and remove the expected errors, because all of these cases should be accepted.
After a bit of DI work, failable initializers should "just work" with a
failure at any point in a class initializer.  It now correctly cleans up any partially or conditionally
initiailized state.  If something is being missed, that is a bad bug that needs to be fixed.


Swift SVN r21624