[g-arc-opts] Guaranteed parameters are always known safe.
If we have an argument that is guaranteed, the optimizer should always treat it
as known safe. This is good to do since we will eventually if I understand
correctly mark self as @guaranteed in methods. For now it will only be exposed
via function signature optimization.

rdar://18289318

Swift SVN r21834