Solver: don't try a default literal type binding when we have contextual type information.
When computing the set of potential bindings for a type variable with
a literal constraint, we suggest the default literal type as a
fallback. Suppress this suggestion in the cases where another
constraint provides a type that already has some kind of contextual
type information (that does conform to the protocol), making the
default literal type more of a fallback. When type-checking the given
declaration:

  var dict: NSDictionary = [
    "status": 200,
    "people": [ [ "id": 255,
                  "name": [ "first": "John", "last": "Appleseed" ] ] ] ]

it reduces the number of solution states explored by 2/3, the number
of attempted type variable bindings by 3/4, and the number of
disjunctions explored by 1/3. This optimization is useful in general,
and lets type annotations guide the type checker to a much greater
extent. It also helps with rdar://problem/18269449.

Swift SVN r21878