Solver: use literal constraints to eagerly look through optional types.
When we have a literal constraint on an optional type, and the
optional does not itself conform to the literal protocol, look through
the optional to determine whether its underlying type conforms to the
protocol. This essentially helps filter out dumb solutions that try
optional types where we should just be relying on the
value-to-optional conversion.

Should be NFC, since we would find the same solutions anyway.

Swift SVN r21918